School Library: Feature/refactor
