+++ classes/app.rb
@@ -60,21 +60,24 @@ class App
   end
 
   def create_rental
-    puts 'Select a book from the following list by number'
-    @books.each_with_index do |book, index|
-      puts "#{index}) Title: \"#{book.title}\", Author: #{book.author}"
-    end
-    book_index = gets.chomp.to_i
-    puts 'Select a person from the following list by number (not id)'
-    @people.each_with_index do |person, index|
-      type = person.instance_of?(Student) ? 'Student' : 'Teacher'
-      puts "#{index}) [#{type}] Name: #{person.name}, ID: #{person.id}, Age: #{person.age}"
-    end
-    person_index = gets.chomp.to_i
+    book_index = select_book
+    person_index = select_person
     print 'Date: '
     date = gets.chomp
     @rentals.push(Rental.new(date, @books[book_index], @people[person_index]))
     puts 'Rental created successfully'
+  end
+
+  def select_book
+    puts 'Select a book from the following list by number:'
+    list_books
+    gets.chomp.to_i
+  end
+
+  def select_person
+    puts 'Select a person from the following list by number (not id):'
+    list_people
+    gets.chomp.to_i
   end
 
   def list_books
